refactor(admin): FeedImageGrid·MobilePhotoGrid 통합 및 desktop/mobile 폴더 정리
